requirements.txt
backports.ssl-match-hostname==3.5.0.1
cffi==1.6.0
cryptography==1.3.2
Django==1.9.6
docker-py==1.8.1
enum34==1.1.5
idna==2.1
ipaddress==1.0.16
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
requests==2.10.0
six==1.10.0
websocket-client==0.37.0
wheel==0.24.0
